Написаны сервисы, мапперы, дто. есть проблема с lombok тесты не видят геттеры

+++ src/main/java/ru/shaineDoc/booksAndAuthors/entity/Author.java
@@ -17,9 +17,9 @@ import java.util.List;
 public class Author {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private LocalDate birthDate;
-    @OneToMany(mappedBy = "author",cascade = CascadeType.ALL, orphanRemoval = true)
+    @OneToMany(mappedBy = "author", cascade = CascadeType.ALL, orphanRemoval = true)
     List<Book> books;
 }
